src/sistema_indicadores/tests/factories.py
import factory

from sistema_indicadores.models import Organization


class OrganizationFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = Organization

    name = factory.Faker("name")
