pyproject.toml
[tool.poetry]
name = "sistema-indicadores"
version = "0.0.0"
description = "Sistema para control de recursos, actividades y muestra de indicadores"
authors = ["Wolfang Torres <wolfang.torres@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
packages = [{include = "sistema_indicadores", from = "src" }]
repository = "https://github.com/WolfangT/sistema_indicadores"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.7"
kpireport = "^0.1.8"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
factory-boy = "^3.2.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"
black = "^23.1.0"
pylint = "^2.16.2"
Sphinx = "^6.1.3"
isort = "^5.12.0"
nose = "^1.3.7"
interrogate = "^1.5.0"
mypy = "<1.0"
refurb = "^1.12.0"
pydocstyle = "^6.3.0"
deptry = "^0.8.0"
flake8 = "^6.0.0"
bandit = "^1.7.4"
django-stubs = {extras = ["compatible-mypy"], version = "^1.14.0"}
pylint-django = "^2.5.3"
pytest-django = "^4.5.2"
django-environ = "^0.9.0"

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

# Plugins

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"

[tool.pytest.ini_options]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
DJANGO_SETTINGS_MODULE = "config.settings"
testpaths = ["tests", "src/*/"]

# Linting and formatign tools

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = """( .eggs | .git | venv | build | dist )"""

[tool.isort]
profile = "black"
src_paths = ["scr"]
skip_glob = [".git", "__pycache__", ".vscode", "*venv", "build", "dist", "*.egg-info"]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
verbose = false
case_sensitive = true
length_sort = false
balanced_wrapping = true
atomic = true
known_django="django"
known_first_party = ["sistema_indicadores", "sistema_indicadores"]
import_heading_stdlib = "Standard Library"
import_heading_django = "django"
import_heading_thirdparty = "Pip"
import_heading_firstparty = "Local"
import_heading_localfolder = "Local"
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pylint.main]
py-version = "3.11"
load-plugins = ["pylint_django", "pylint_django.checkers.migrations"]

[tool.pylint.format]
max-line-length = "88"

[tool.pylint.reports]
output-format="text"
msg-template="{path}:{module}:{line}: [{msg_id}({symbol}), {obj}] {msg}"

[tool.pylint.logging]
logging-format-style="new"
logging-modules="logging"

[tool.pylint."django foreign keys referenced by strings"]
django-settings-module = "config.settings"

[tool.refurb]
enable_all = true
python-version = 3.11

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]
exclude = '/(venv|instance|test|\..*)/$'

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.bandit]
exclude_dirs = ["tests", "venv"]
